added the task color properties inside the 'Task' data model
