+++ ManageTask/Home/Views/HomeView.swift
@@ -10,26 +10,12 @@ import SwiftUI
 struct HomeView: View {
     @StateObject var viewModel = HomeViewModel()
     
-    func tasksWithStatus(_ completionStatus: TaskCompletionStatus) -> [Binding<Task>] {
-        switch completionStatus {
-        case .overdue:
-            return $viewModel.allTasks.filter{ $0.completionDate.wrappedValue == nil && $0.dueDate.wrappedValue < Date()}
-            
-        case .incomplete:
-            return $viewModel.allTasks.filter{ $0.completionDate.wrappedValue == nil && $0.dueDate.wrappedValue > Date() }
-            
-        case .completed:
-            return $viewModel.allTasks.filter{ $0.completionDate.wrappedValue != nil}
-        }
-        
-    }
-    
     var body: some View {
         NavigationStack {
             List {
                 ForEach(TaskCompletionStatus.allCases, id: \.hashValue) { completionStatus in
                     Section(completionStatus.rawValue) {
-                        ForEach(tasksWithStatus(completionStatus)) { task in
+                        ForEach(viewModel.tasksWithStatus(completionStatus)) { task in
                             ZStack(alignment: .leading) {
                                 TaskView(task: task)
                                 NavigationLink {
