+++ ManageTask/Models/Task.swift
@@ -17,6 +17,30 @@ enum TaskPriority: Int {
     }
 }
 
+enum TaskColorString: String, CaseIterable {
+    case red, orange, green, mint, teal, cyan, blue, indigo, purple, pink, brown
+}
+
+extension Color {
+    init?(_ taskColor: TaskColorString?) {
+        switch taskColor?.rawValue {
+        case "red": self = .red
+        case "orange": self = .orange
+        case "yellow": self = .yellow
+        case "green": self = .green
+        case "mint": self = .mint
+        case "teal": self = .teal
+        case "cyan": self = .cyan
+        case "blue": self = .blue
+        case "indigo": self = .indigo
+        case "purple": self = .purple
+        case "pink": self = .pink
+        case "brown": self = .brown
+        default: return nil
+        }
+    }
+}
+
 
 struct Task: Identifiable {
     let id = UUID()
@@ -25,6 +49,7 @@ struct Task: Identifiable {
     var priority: TaskPriority
     var notes: String?
     var completionDate: Date? = nil
+    let taskColor = Color(TaskColorString.allCases.randomElement())
     
     init(title: String, dueDate: Date, priority: TaskPriority, notes: String? = nil, completionDate: Date? = nil) {
         self.title = title
